+++ changelog.md
@@ -7,6 +7,7 @@
 * LLT-3468: Separate Heartbeat collection from QoS so heartbeat collection can be alone enabled.
 * LLT-3502: Refactor telio-nurse. Fix meshnet members list, conectivity matrix, external links data.
 * LLT-3504: Enable nurse component. Its start has been accidentally removed during merges.
+* LLT-3562: Add offline external nodes in analytics events.
 
 ### v4.0.0
 ### **Bruschetta**